Move ModelViewer to ExampleBrowser
