--- Source/Examples/WPF/ModelViewer/App.xaml
@@ -1,8 +1,0 @@
-﻿<Application x:Class="StudioDemo.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    StartupUri="Window1.xaml">
-    <Application.Resources>
-         
-    </Application.Resources>
-</Application>
