--- Source/Examples/WPF/ModelViewer/App.xaml.cs
@@ -1,20 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="App.xaml.cs" company="Helix Toolkit">
-//   Copyright (c) 2014 Helix Toolkit contributors
-// </copyright>
-// <summary>
-//   Interaction logic for App.xaml
-// </summary>
-// --------------------------------------------------------------------------------------------------------------------
-
-namespace ModelViewer
-{
-    using System.Windows;
-
-    /// <summary>
-    /// Interaction logic for App.xaml
-    /// </summary>
-    public partial class App : Application
-    {
-    }
-}
