--- Source/Examples/WPF/ModelViewer/Observable.cs
@@ -1,24 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="Observable.cs" company="Helix Toolkit">
-//   Copyright (c) 2014 Helix Toolkit contributors
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-using System.ComponentModel;
-
-namespace ModelViewer
-{
-    public class Observable : INotifyPropertyChanged
-    {
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        protected void RaisePropertyChanged(string property)
-        {
-            var handler = PropertyChanged;
-            if (handler != null)
-            {
-                handler(this, new PropertyChangedEventArgs(property));
-            }
-        }
-    }
-}
