--- Source/Examples/WPF/ModelViewer/Properties/AssemblyInfo.cs
@@ -1,13 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="AssemblyInfo.cs" company="Helix Toolkit">
-//   Copyright (c) 2014 Helix Toolkit contributors
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-using System.Reflection;
-using System.Windows;
-
-[assembly: AssemblyTitle("Helix 3D Toolkit model viewer demo")]
-[assembly: AssemblyDescription("")]
-
-[assembly: ThemeInfo(ResourceDictionaryLocation.None, ResourceDictionaryLocation.SourceAssembly)]
