--- Source/Examples/WPF/ModelViewer/Window1.xaml
@@ -1,100 +1,0 @@
-﻿<Window x:Class="ModelViewer.Window1"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:helix="clr-namespace:HelixToolkit.Wpf;assembly=HelixToolkit.Wpf"
-        xmlns:modelViewer="clr-namespace:ModelViewer" Title="{Binding ApplicationTitle}" Height="480" Width="640" Background="Black">
-    <Window.Resources>
-        <!--   <HierarchicalDataTemplate DataType = "{x:Type media3d:ModelVisual3D}" ItemsSource = "{Binding Children}">
-            <StackPanel Orientation="Horizontal">
-                <Rectangle Width="8" Height="8" Fill="Red" Stroke="Black"/>
-                <TextBlock Text="{Binding}"/>
-            </StackPanel>
-        </HierarchicalDataTemplate>
-        <HierarchicalDataTemplate DataType = "{x:Type media3d:Model3DGroup}" ItemsSource = "{Binding Children}">
-            <StackPanel Orientation="Horizontal">
-                <Rectangle Width="8" Height="8" Fill="Yellow" Stroke="Black"/>
-                <TextBlock Text="{Binding}"/>
-            </StackPanel>
-        </HierarchicalDataTemplate>
-           <HierarchicalDataTemplate DataType = "{x:Type media3d:Model3DGroup}" ItemsSource = "{Binding Path=Children}">
-                <StackPanel Orientation="Horizontal">
-                <Rectangle Width="8" Height="8" Fill="Yellow" Stroke="Black"/>
-                <TextBlock Text="{Binding}"/>
-            <ItemsControl ItemsSource="{Binding Path=Children}">           
-            </ItemsControl>
-            </StackPanel>
-
-        </HierarchicalDataTemplate>-->
-        <HierarchicalDataTemplate DataType = "{x:Type modelViewer:VisualViewModel}" ItemsSource = "{Binding Children}">
-            <StackPanel Orientation="Vertical">
-                <StackPanel Orientation="Horizontal">
-                    <Ellipse Width="8" Height="8" Margin="4" Fill="{Binding Brush}" Stroke="Black"/>
-                    <TextBlock Text="{Binding TypeName}" FontWeight="Bold"/>
-                    <TextBlock Text="{Binding Name}" Padding="8 0 0 0"/>
-                </StackPanel>
-            </StackPanel>
-        </HierarchicalDataTemplate>
-        <!--<HierarchicalDataTemplate DataType = "{x:Type media3d:ModelVisual3D}" ItemsSource = "{Binding Children}">
-            <StackPanel Orientation="Vertical">
-            <StackPanel Orientation="Horizontal">
-                <Ellipse Width="8" Height="8" Margin="4" Fill="Blue" Stroke="Black"/>
-                <TextBlock Text="{Binding}"/>
-                <TextBlock Text="{Binding Visual3DModel}"/>
-            </StackPanel>
-                <ContentControl Content="{Binding Content}"/>
-            </StackPanel>
-        </HierarchicalDataTemplate>
-        <HierarchicalDataTemplate DataType = "{x:Type media3d:Model3DGroup}" ItemsSource="{Binding Children}">
-                <StackPanel Orientation="Horizontal">
-                    <Ellipse Width="8" Height="8" Margin="4" Fill="Green" Stroke="Black"/>
-                    <TextBlock Text="{Binding}"/>
-                </StackPanel>
-        </HierarchicalDataTemplate>
-        <DataTemplate DataType = "{x:Type media3d:GeometryModel3D}">
-            <StackPanel Orientation="Horizontal">
-                <Ellipse Width="8" Height="8" Margin="4" Fill="Red" Stroke="Black"/>
-                <TextBlock Text="{Binding}"/>
-            </StackPanel>
-        </DataTemplate>-->
-    </Window.Resources>
-    <DockPanel>
-        <Menu DockPanel.Dock="Top">
-            <MenuItem Header="_File">
-                <MenuItem Header="_Open..." Command="{Binding FileOpenCommand}"/>
-                <MenuItem Header="_Export..." Command="{Binding FileExportCommand}"/>
-                <Separator/>
-                <MenuItem Header="E_xit" Command="{Binding FileExitCommand}"/>
-            </MenuItem>
-            <MenuItem Header="_Edit">
-                <MenuItem Header="_Copy XAML" Command="{Binding EditCopyXamlCommand}"/>
-            </MenuItem>
-            <MenuItem Header="_View">
-                <MenuItem Header="_Zoom extents" Command="{Binding ViewZoomExtentsCommand}"/>
-            </MenuItem>
-            <MenuItem Header="_Help">
-                <MenuItem Header="_About" Command="{Binding HelpAboutCommand}"/>
-            </MenuItem>
-        </Menu>
-
-        <DockPanel DockPanel.Dock="Left" Width="400" >
-            <Slider Value="{Binding Expansion}" Minimum="1" Maximum="2" DockPanel.Dock="Bottom"/>
-            <TextBox DockPanel.Dock="Bottom" Height="200" Text="{Binding Info}"/>
-            <TreeView ItemsSource="{Binding Elements}">
-            </TreeView>
-        </DockPanel>
-
-        <helix:HelixViewport3D x:Name="view1" CameraRotationMode="Trackball" ModelUpDirection="0,1,0">
-            <ModelVisual3D x:Name="root1">
-                <!--<helix:CubeVisual3D Fill="Red"/> 
-                <helix:GridLinesVisual3D/>-->
-                <helix:DefaultLights/>
-                <ModelVisual3D Content="{Binding CurrentModel}"/>
-                <!--<helix:FileModelVisual3D Source="{Binding CurrentModelPath, Mode=TwoWay}" ModelLoaded="ModelLoaded"/>-->
-                <!--<helix:Expander3D Expansion="{Binding Expansion}" Content="{Binding CurrentModel}"/>-->
-                <!-- <helix:HelixViewport3D.Triggers>
-                <EventTrigger 
-            </helix:HelixViewport3D.Triggers>-->
-            </ModelVisual3D>
-        </helix:HelixViewport3D>
-    </DockPanel>
-</Window>
