+++ Source/HelixToolkit.Wpf.sln
@@ -25,8 +25,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BatchRender", "Tools\BatchR
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MvvmDemo", "Examples\WPF\MvvmDemo\MvvmDemo.csproj", "{47FC8A4C-86B6-4D9D-BE10-EE74F8B992BB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ModelViewer", "Examples\WPF\ModelViewer\ModelViewer.csproj", "{AB6A4302-6B39-4606-874C-53CE263BB2A2}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleBrowser", "Examples\WPF\ExampleBrowser\ExampleBrowser.csproj", "{991A4C2B-D830-41CB-A810-047ADCF69219}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Wpf", "Wpf", "{848BE1AD-560A-4419-BD41-994A1A1CCF1F}"
@@ -141,16 +139,6 @@ Global
 		{47FC8A4C-86B6-4D9D-BE10-EE74F8B992BB}.Release|Mixed Platforms.Build.0 = Release|x86
 		{47FC8A4C-86B6-4D9D-BE10-EE74F8B992BB}.Release|x86.ActiveCfg = Release|x86
 		{47FC8A4C-86B6-4D9D-BE10-EE74F8B992BB}.Release|x86.Build.0 = Release|x86
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2}.Release|x86.ActiveCfg = Release|Any CPU
 		{991A4C2B-D830-41CB-A810-047ADCF69219}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{991A4C2B-D830-41CB-A810-047ADCF69219}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{991A4C2B-D830-41CB-A810-047ADCF69219}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -212,7 +200,6 @@ Global
 		{DFE51472-60CD-4F9C-8E52-47B250F58850} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC} = {4EC16C60-5AFA-4854-82D9-FB1BF92C3812}
 		{47FC8A4C-86B6-4D9D-BE10-EE74F8B992BB} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
-		{AB6A4302-6B39-4606-874C-53CE263BB2A2} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{991A4C2B-D830-41CB-A810-047ADCF69219} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{848BE1AD-560A-4419-BD41-994A1A1CCF1F} = {A710CED9-33A6-42F1-8A09-55FEC57B6460}
 		{6C9BF028-EDF1-49AC-B441-B886DD807E04} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
